+++ osmnx/geocoder.py
@@ -29,7 +29,7 @@ def geocode(query):
     params = OrderedDict()
     params["format"] = "json"
     params["limit"] = 1
-    params["dedupe"] = 0  # prevent OSM deduping results so we get precisely 'limit' # of results
+    params["dedupe"] = 0  # prevent deduping to get precise number of results
     params["q"] = query
     response_json = downloader.nominatim_request(params=params)
 
@@ -44,14 +44,24 @@ def geocode(query):
         raise Exception(f'Nominatim geocoder returned no results for query "{query}"')
 
 
-def geocode_to_gdf(query, which_result=None, buffer_dist=None):
+def geocode_to_gdf(query, which_result=None, by_osmid=False, buffer_dist=None):
     """
-    Geocode a query or queries to a GeoDataFrame with the Nominatim geocoder.
+    Retrieve place(s) by name or ID from the Nominatim API as a GeoDataFrame.
 
-    Geometry column contains place boundaries if they exist in OpenStreetMap.
-    Query can be a string or dict, or a list of strings/dicts to send to the
-    geocoder. If query is a list, then which_result should be either a single
-    value or a list of the same length as query.
+    You can query by place name or OSM ID. If querying by place name, the
+    query argument can be a string or structured dict, or a list of such
+    strings/dicts to send to geocoder. You can instead query by OSM ID by
+    setting `by_osmid=True`. In this case, geocode_to_gdf treats the query
+    argument as an OSM ID (or list of OSM IDs) for Nominatim lookup rather
+    than text search. OSM IDs must be prepended with their types: node (N),
+    way (W), or relation (R), in accordance with the Nominatim format. For
+    example, `query='["R2192363", "N240109189", "W427818536"]'`.
+
+    If query argument is a list, then which_result should be either a single
+    value or a list with the same length as query. The queries you provide
+    must be resolvable to places in the Nominatim database. The resulting
+    GeoDataFrame's geometry column contains place boundaries if they exist in
+    OpenStreetMap.
 
     Parameters
     ----------
@@ -59,7 +69,10 @@ def geocode_to_gdf(query, which_result=None, buffer_dist=None):
         query string(s) or structured dict(s) to geocode
     which_result : int
         which geocoding result to use. if None, auto-select the first
-        (Multi)Polygon or raise an error if OSM doesn't return one.
+        (Multi)Polygon or raise an error if OSM doesn't return one. to get
+        the top match regardless of geometry type, set which_result=1
+    by_osmid : bool
+        if True, handle query as an OSM ID for lookup rather than text search
     buffer_dist : float
         distance to buffer around the place geometry, in meters
 
@@ -94,7 +107,7 @@ def geocode_to_gdf(query, which_result=None, buffer_dist=None):
     # geocode each query and add to GeoDataFrame as a new row
     gdf = gpd.GeoDataFrame()
     for q, wr in zip(query, which_result):
-        gdf = gdf.append(_geocode_query_to_gdf(q, wr))
+        gdf = gdf.append(_geocode_query_to_gdf(q, wr, by_osmid))
 
     # reset GeoDataFrame index and set its CRS
     gdf = gdf.reset_index(drop=True)
@@ -112,7 +125,7 @@ def geocode_to_gdf(query, which_result=None, buffer_dist=None):
     return gdf
 
 
-def _geocode_query_to_gdf(query, which_result):
+def _geocode_query_to_gdf(query, which_result, by_osmid):
     """
     Geocode a single place query to a GeoDataFrame.
 
@@ -122,7 +135,10 @@ def _geocode_query_to_gdf(query, which_result):
         query string or structured dict to geocode
     which_result : int
         which geocoding result to use. if None, auto-select the first
-        (Multi)Polygon or raise an error if OSM doesn't return one.
+        (Multi)Polygon or raise an error if OSM doesn't return one. to get
+        the top match regardless of geometry type, set which_result=1
+    by_osmid : bool
+        if True, handle query as an OSM ID for lookup rather than text search
 
     Returns
     -------
@@ -134,12 +150,16 @@ def _geocode_query_to_gdf(query, which_result):
     else:
         limit = which_result
 
-    results = downloader._osm_polygon_download(query, limit=limit)
+    results = downloader._osm_place_download(query, by_osmid=by_osmid, limit=limit)
 
     # choose the right result from the JSON response
     if not results:
         # if no results were returned, raise error
         raise ValueError(f'OSM returned no results for query "{query}"')
+
+    elif by_osmid:
+        # if searching by OSM ID, always take the first (ie, only) result
+        result = results[0]
 
     elif which_result is None:
         # else, if which_result=None, auto-select the first (Multi)Polygon
@@ -154,31 +174,35 @@ def _geocode_query_to_gdf(query, which_result):
         msg = f'OSM returned fewer than `which_result={which_result}` results for query "{query}"'
         raise ValueError(msg)
 
-    # build the geojson feature from the chosen result
-    south, north, west, east = [float(x) for x in result["boundingbox"]]
-    place = result["display_name"]
-    geometry = result["geojson"]
-    features = [
-        {
-            "type": "Feature",
-            "geometry": geometry,
-            "properties": {
-                "place_name": place,
-                "bbox_north": north,
-                "bbox_south": south,
-                "bbox_east": east,
-                "bbox_west": west,
-            },
-        }
-    ]
-
     # if we got a non (Multi)Polygon geometry type (like a point), log warning
-    if geometry["type"] not in {"Polygon", "MultiPolygon"}:
-        msg = f'OSM returned a {geometry["type"]} as the geometry for query "{query}"'
+    geom_type = result["geojson"]["type"]
+    if geom_type not in {"Polygon", "MultiPolygon"}:
+        msg = f'OSM returned a {geom_type} as the geometry for query "{query}"'
         utils.log(msg, level=lg.WARNING)
 
+    # build the GeoJSON feature from the chosen result
+    south, north, west, east = result["boundingbox"]
+    feature = {
+        "type": "Feature",
+        "geometry": result["geojson"],
+        "properties": {
+            "bbox_north": north,
+            "bbox_south": south,
+            "bbox_east": east,
+            "bbox_west": west,
+        },
+    }
+
+    # add the other attributes we retrieved
+    for attr in result:
+        if attr not in {"address", "boundingbox", "geojson", "icon", "licence"}:
+            feature["properties"][attr] = result[attr]
+
     # create and return the GeoDataFrame
-    return gpd.GeoDataFrame.from_features(features)
+    gdf = gpd.GeoDataFrame.from_features([feature])
+    cols = ["lat", "lon", "bbox_north", "bbox_south", "bbox_east", "bbox_west"]
+    gdf[cols] = gdf[cols].astype(float)
+    return gdf
 
 
 def _get_first_polygon(results, query):
@@ -188,7 +212,7 @@ def _get_first_polygon(results, query):
     Parameters
     ----------
     results : list
-        list of results from downloader._osm_polygon_download
+        list of results from downloader._osm_place_download
     query : str
         the query string or structured dict that was geocoded